Add Linux Implementation
